pyproject.toml
[project]
name = "goit-pythonweb-hw-12"
version = "0.1.0"
description = ""
authors = [
    {name = "Max Samsonenko",email = "samsonenko.maxwork@gmail.com"}
]
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi[standard] (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.0,<0.35.0)",
    "sqlalchemy (>=2.0.40,<3.0.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "pydantic (>=2.11.2,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "python-jose[cryptography] (>=3.4.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "libgravatar (>=1.0.4,<2.0.0)",
    "pydantic-settings (>=2.8.1,<3.0.0)",
    "slowapi (>=0.1.9,<0.2.0)",
    "fastapi-mail (>=1.4.2,<2.0.0)",
    "cloudinary (>=1.43.0,<2.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "pytest-asyncio (>=0.26.0,<0.27.0)",
    "aiosqlite (>=0.21.0,<0.22.0)",
    "redis (>=5.2.1,<6.0.0)",
    "redis-lru (>=0.1.2,<0.2.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
sphinx = "^8.2.3"
pytest-cov = "^6.1.1"

